Move settingsStore to utils
